+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/aio/_text_analytics_client_async.py
@@ -47,6 +47,8 @@ from .._models import (
     RecognizeLinkedEntitiesAction,
     AnalyzeSentimentAction,
     AnalyzeHealthcareEntitiesResult,
+    ExtractSummaryAction,
+    ExtractSummaryResult
 )
 from .._lro import TextAnalyticsOperationResourcePolling
 from ._lro_async import (
@@ -848,6 +850,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                 RecognizePiiEntitiesAction,
                 ExtractKeyPhrasesAction,
                 AnalyzeSentimentAction,
+                ExtractSummaryAction
             ]
         ],  # pylint: disable=line-too-long
         **kwargs: Any,
@@ -860,6 +863,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                     RecognizePiiEntitiesResult,
                     ExtractKeyPhrasesResult,
                     AnalyzeSentimentResult,
+                    ExtractSummaryResult,
                     DocumentError,
                 ]
             ]
@@ -885,7 +889,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             Duplicate actions in list not supported.
         :type actions:
             list[RecognizeEntitiesAction or RecognizePiiEntitiesAction or ExtractKeyPhrasesAction or
-            RecognizeLinkedEntitiesAction or AnalyzeSentimentAction]
+            RecognizeLinkedEntitiesAction or AnalyzeSentimentAction, or ExtractSummaryAction]
         :keyword str display_name: An optional display name to set for the requested analysis.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
@@ -909,8 +913,13 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         :rtype:
             ~azure.ai.textanalytics.aio.AsyncAnalyzeActionsLROPoller[~azure.core.async_paging.AsyncItemPaged[
             list[Union[RecognizeEntitiesResult, RecognizeLinkedEntitiesResult, RecognizePiiEntitiesResult,
-            ExtractKeyPhrasesResult, AnalyzeSentimentResult, DocumentError]]]]
+            ExtractKeyPhrasesResult, AnalyzeSentimentResult, ExtractSummaryAction, DocumentError]]]]
         :raises ~azure.core.exceptions.HttpResponseError or TypeError or ValueError or NotImplementedError:
+
+        .. versionadded:: v3.1
+            The *begin_analyze_actions* client method.
+        .. versionadded:: v3.2-preview
+            The *ExtractSummaryAction* input option and *ExtractSummaryResult* result object
 
         .. admonition:: Example:
 
@@ -985,6 +994,14 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                         == _AnalyzeActionsType.ANALYZE_SENTIMENT
                     ]
                 ],
+                extractive_summarization_tasks=[
+                    t._to_generated()  # pylint: disable=protected-access
+                    for t in [
+                        a
+                        for a in actions
+                        if _determine_action_type(a) == _AnalyzeActionsType.EXTRACT_SUMMARY
+                    ]
+                ]
             )
             analyze_body = self._client.models(api_version=self._api_version).AnalyzeBatchInput(
                 display_name=display_name, tasks=analyze_tasks, analysis_input=docs