+++ sdk/textanalytics/azure-ai-textanalytics/samples/async_samples/sample_analyze_actions_async.py
@@ -12,8 +12,8 @@ FILE: sample_analyze_actions_async.py
 DESCRIPTION:
     This sample demonstrates how to submit a collection of text documents for analysis, which consists of a variety
     of text analysis actions, such as Entity Recognition, PII Entity Recognition, Linked Entity Recognition,
-    Sentiment Analysis, or Key Phrase Extraction.  The response will contain results from each of the individual
-    actions specified in the request.
+    Sentiment Analysis, Key Phrase Extraction, or Extractive Text Summarization.  The response will contain results
+    from each of the individual actions specified in the request.
 
 USAGE:
     python sample_analyze_actions_async.py
@@ -38,6 +38,7 @@ async def sample_analyze_async():
         RecognizePiiEntitiesAction,
         ExtractKeyPhrasesAction,
         AnalyzeSentimentAction,
+        ExtractSummaryAction,
     )
 
     endpoint = os.environ["AZURE_TEXT_ANALYTICS_ENDPOINT"]
@@ -49,15 +50,15 @@ async def sample_analyze_async():
     )
 
     documents = [
-        'We went to Contoso Steakhouse located at midtown NYC last week for a dinner party, and we adore the spot!'\
-        'They provide marvelous food and they have a great menu. The chief cook happens to be the owner (I think his name is John Doe)'\
-        'and he is super nice, coming out of the kitchen and greeted us all.'\
+        'We went to Contoso Steakhouse located at midtown NYC last week for a dinner party, and we adore the spot! '
+        'They provide marvelous food and they have a great menu. The chief cook happens to be the owner (I think his name is John Doe) '
+        'and he is super nice, coming out of the kitchen and greeted us all.'
         ,
 
-        'We enjoyed very much dining in the place!'\
-        'The Sirloin steak I ordered was tender and juicy, and the place was impeccably clean. You can even pre-order from their'\
-        'online menu at www.contososteakhouse.com, call 312-555-0176 or send email to order@contososteakhouse.com!'\
-        'The only complaint I have is the food didn\'t come fast enough. Overall I highly recommend it!'\
+        'We enjoyed very much dining in the place! '
+        'The Sirloin steak I ordered was tender and juicy, and the place was impeccably clean. You can even pre-order from their '
+        'online menu at www.contososteakhouse.com, call 312-555-0176 or send email to order@contososteakhouse.com! '
+        'The only complaint I have is the food didn\'t come fast enough. Overall I highly recommend it!'
     ]
 
     async with text_analytics_client:
@@ -69,7 +70,8 @@ async def sample_analyze_async():
                 RecognizePiiEntitiesAction(),
                 ExtractKeyPhrasesAction(),
                 RecognizeLinkedEntitiesAction(),
-                AnalyzeSentimentAction()
+                AnalyzeSentimentAction(),
+                ExtractSummaryAction()
             ]
         )
 
@@ -153,6 +155,17 @@ async def sample_analyze_async():
                     analyze_sentiment_result.confidence_scores.negative,
                 ))
             print("------------------------------------------")
+
+            extract_summary_result = action_results[5]
+            print("...Results of Extractive Text Summarization action:")
+            if extract_summary_result.is_error:
+                print("...Is an error with code '{}' and message '{}'".format(
+                    extract_summary_result.code, extract_summary_result.message
+                ))
+            else:
+                print("Summary extracted: {}".format(
+                    " ".join([sentence.text for sentence in extract_summary_result.sentences]))
+                )
 
     # [END analyze_async]