+++ sdk/textanalytics/azure-ai-textanalytics/CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## 5.2.0b1 (Unreleased)
 
+This version of the SDK defaults to the latest supported API version, which currently is `v3.2-preview.1`.
+
 ### Features Added
+- Added support for Extractive Summarization actions through the `ExtractSummaryAction` type.
 
 ### Breaking Changes
 
 ### Bugs Fixed
+- `RecognizePiiEntitiesAction` option `disable_service_logs` now correctly defaults to `True`.
 
 ### Other Changes